Retry varg inference after inferring targs
Basically, the idea here is to give method arguments
a second chance to infer their own type arguments once
the outer call has inferred all of its type arguments.
This allows us to do a much better job of inference
when a polymorphic method is being turned into a function.